check succeeds now
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/clusterds@306 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/streamFPM/src/Node.h
@@ -9,15 +9,15 @@ public:
     int content() { return mContent; }
     void setContent(int c) { mContent = c; }
 
-    void setCount(int cnt) { count = cnt; }
-    int getCount() { return count; }
-    void incrementCount(int add = 1){ count = count + add; }
+    void setCount(double cnt) { count = cnt; }
+    double getCount() { return count; }
+    void incrementCount(double add = 1){ count = count + add; }
     
     Node* findChild(int c);
     void appendChild(Node* child) { mChildren.push_back(child); }
     std::vector<Node*> children() { return mChildren; }
-    void setErr(int e) { err = e; }
-    int getErr() { return err; }
+    void setErr(double e) { err = e; }
+    double getErr() { return err; }
     void setId(int i) { id = i; }
     int getId() { return id; }
     
@@ -35,8 +35,8 @@ private:
     //mContent is the item that the node is tracking
     int mContent;
     bool mMarker;
-    int count;
-    int err;
+    double count;
+    double err;
     //id = TID last transaction that had this item
     int id;
